+++ vector/src/main/java/im/vector/app/core/extensions/FragmentManager.kt
@@ -17,12 +17,12 @@
 package im.vector.app.core.extensions
 
 import androidx.fragment.app.FragmentTransaction
-import org.matrix.android.sdk.api.extensions.tryThis
+import org.matrix.android.sdk.api.extensions.tryOrNull
 
 inline fun androidx.fragment.app.FragmentManager.commitTransactionNow(func: FragmentTransaction.() -> FragmentTransaction) {
     // Could throw and make the app crash
     // e.g sharedActionViewModel.observe()
-    tryThis("Failed to commitTransactionNow") {
+    tryOrNull("Failed to commitTransactionNow") {
         beginTransaction().func().commitNow()
     }
 }