+++ vector/src/main/java/im/vector/app/features/notifications/NotificationUtils.kt
@@ -775,7 +775,7 @@ class NotificationUtils @Inject constructor(private val context: Context,
     /**
      * Build the summary notification
      */
-    fun buildSummaryListNotification(style: NotificationCompat.InboxStyle,
+    fun buildSummaryListNotification(style: NotificationCompat.InboxStyle?,
                                      compatSummary: String,
                                      noisy: Boolean,
                                      lastMessageTimestamp: Long): Notification {